requirements/dev.txt
black
django-stubs[compatible-mypy]
flake8
flake8-django
pylint
rope
twine
types-requests
types-tzlocal
mypy
